CircleSeeker_2.0.1/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "circleseeker"
version = "2.0.1"
description = "Comprehensive eccDNA detection from HiFi sequencing data"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-2.0"}
authors = [
    {name = "CircleSeeker Team", email = "circleseeker@example.org"}
]
maintainers = [
    {name = "CircleSeeker Team", email = "circleseeker@example.org"}
]
keywords = [
    "bioinformatics",
    "eccDNA",
    "circular-dna",
    "pacbio",
    "hifi-sequencing",
    "genomics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]

dependencies = [
    "click>=8.1",
    "pandas>=2.0",
    "numpy>=1.23",
    "biopython>=1.81",
    "pysam>=0.22",
    "pyyaml>=6.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.3",
    "black>=23.7",
    "isort>=5.12",
    "flake8>=6.1",
    "mypy>=1.5",
    "pre-commit>=3.3",
]
docs = [
    "sphinx>=7.1",
    "sphinx-rtd-theme>=1.3",
    "sphinx-click>=5.0",
    "myst-parser>=2.0",
]
test = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "hypothesis>=6.82",
]

[project.urls]
Homepage = "https://github.com/circleseeker/circleseeker2"
Documentation = "https://circleseeker2.readthedocs.io"
Repository = "https://github.com/circleseeker/circleseeker2"
Issues = "https://github.com/circleseeker/circleseeker2/issues"
Changelog = "https://github.com/circleseeker/circleseeker2/blob/main/CHANGELOG.md"

[project.scripts]
circleseeker = "circleseeker.cli:main"
CircleSeeker = "circleseeker.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
circleseeker = [
    "resources/*.yaml",
    "reports/templates/*.j2",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=circleseeker",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["circleseeker"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
